Fix for the module version number.
Module version numbers came from the configure scripts but the move to
one script builds - the modules build - broke that because those
configure scripts are not run anymore. I added two Makefile variables
to Makefile.am to hold the module name and version; a less than
optimal solution since now the information has to be maintained in two
or three places (configure.ac and .spec for the standard build and
Makefile.am for the modules build).